refactor: teleporter-bundle -> teleporter
