+++ src/items/systems/teleporter_system.rs
@@ -1,8 +1,8 @@
+use crate::items::events::OnEnterEvent;
 use crate::items::primitive::enterable::EnterAble;
 use crate::items::primitive::is_usable::IsUsable;
 use crate::items::primitive::player_only::PersonOnly;
-use crate::items::primitive::teleporter::Teleporter;
-use crate::items::events::OnEnterEvent;
+use crate::items::teleporter::Teleporter;
 use crate::player::components::player::Person;
 use bevy::prelude::*;
 use bevy::utils::HashSet;
